Fix build failure - UNTESTED ON ACTUAL HARDWARE
The package version numbering of the Armbian uboot that we are using
continues to be unrelated to the version of uboot that they have
packaged.

This simply updates to the "version number" that is available at
https://us.mirrors.fossho.st/armbian/apt/pool/main/l/linux-u-boot-bananapi-current/

The bananapi hardware is not currently available for testing, so this
is purely a build-time fix, with no testing.